Game Manager Implementation
This branch has a good implementation of a Game Manager that is a singleton class (i.e. only one of them can be made), static (so the variables and methods of the GM can be accessed from anywhere by just using its name as a reference), and is persistent (so it will stay in all new scenes loaded). Tarodev explains. https://www.youtube.com/watch?v=tE1qH8OxO2Y